+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - add events
 - add users
 - add alert sources
+- add escalation policy